+++ spec/contentful/webhook/listener/support/null_logger_spec.rb
@@ -1,0 +1,17 @@
+require 'spec_helper'
+
+describe Contentful::Webhook::Listener::Support::NullLogger do
+  let(:logger) { Contentful::Webhook::Listener::Support::NullLogger.new }
+
+  describe "instance methods" do
+    [:info, :debug, :warn, :error, :write].each do |name|
+      it "##{name} returns nil" do
+        expect(logger.send(name, "foo")).to eq nil
+      end
+
+      it "##{name} doesn't output to STDOUT" do
+        expect { logger.send(name, "foo") }.to_not output("foo").to_stdout
+      end
+    end
+  end
+end